requirements.txt
boto3>=1.26.0
click>=8.0.0
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.0.0
gunicorn>=20.1.0
mysqlclient>=2.1.0
passlib>=1.7.4
psycopg2-binary>=2.9.0
psutil>=5.9.0
pyotp>=2.8.0
python-dotenv>=0.20.0
qrcode>=7.3.1
requests>=2.28.0
SQLAlchemy>=2.0.0
Werkzeug>=2.0.0
WTForms>=3.0.0
